Extract product description and features from license bundle card description component for a reusable copies.
